+++ src/dictionary/SelectEdit.ts
@@ -1,9 +1,8 @@
-import PaginationData, { PaginationDataInitOption } from "../module/PaginationData"
-import DefaultLoadEdit, { DefaultLoadEditInitOption } from "./DefaultLoadEdit"
+import PaginationData from "../module/PaginationData"
+import DefaultSelectEdit, { DefaultSelectEditInitOption } from "./DefaultSelectEdit"
 import DictionaryValue from "../lib/DictionaryValue"
-import SelectValue, { DefaultSelectValueType, SelectValueInitOption, SelectValueType } from "../lib/SelectValue"
-import CascaderValue, { CascaderValueInitOption, CascaderValueType, DefaultCascaderValueType } from "../lib/CascaderValue"
-import SelectData from "../core/SelectData"
+import { DefaultSelectValueType, SelectValueType } from "../lib/SelectValue"
+import { CascaderValueType, DefaultCascaderValueType } from "../lib/CascaderValue"
 
 export interface SelectEditOption {
   hideArrow?: boolean
@@ -24,56 +23,26 @@ export interface searchOption extends searchInitOption {
   value: undefined | string
 }
 
-export interface SelectEditInitOption<C extends PropertyKey | undefined = undefined, D extends (C extends PropertyKey ? CascaderValueType<C> : SelectValueType) = (C extends PropertyKey ? DefaultCascaderValueType<C> : DefaultSelectValueType)> extends DefaultLoadEditInitOption {
+export interface SelectEditInitOption<C extends PropertyKey | undefined = undefined, D extends (C extends PropertyKey ? CascaderValueType<C> : SelectValueType) = (C extends PropertyKey ? DefaultCascaderValueType<C> : DefaultSelectValueType)> extends DefaultSelectEditInitOption<C, D> {
   type: C extends undefined ? 'select' : 'cascader'
-  cascader: C
-  select?: C extends undefined ? (SelectValueInitOption<D> | SelectValue<D>) : (CascaderValueInitOption<C, D> | CascaderValue<C, D>)
   option?: Partial<SelectEditOption>
   search?: searchInitOption
-  pagination?: PaginationDataInitOption | PaginationData
 }
 
-class SelectEdit<C extends PropertyKey | undefined = undefined, D extends (C extends PropertyKey ? CascaderValueType<C> : SelectValueType) = (C extends PropertyKey ? DefaultCascaderValueType<C> : DefaultSelectValueType)> extends DefaultLoadEdit {
+class SelectEdit<C extends PropertyKey | undefined = undefined, D extends (C extends PropertyKey ? CascaderValueType<C> : SelectValueType) = (C extends PropertyKey ? DefaultCascaderValueType<C> : DefaultSelectValueType)> extends DefaultSelectEdit<C, D> {
   static $name = 'SelectEdit'
-  static $defaultPlaceholder = function (name: string) {
-    return `请选择${name}`
-  }
   static $defaultOption = {
     hideArrow: false,
     hideClear: false,
     autoWidth: false
   }
   type: C extends undefined ? 'select' : 'cascader'
-  cascader: C
-  $select: C extends undefined ? SelectValue<D> : CascaderValue<C, D>
   $option: SelectEditOption
   $search?: searchOption
   $pagination?: PaginationData
   constructor(initOption: SelectEditInitOption<C>, parent?: DictionaryValue, modName?: string) {
-    if (initOption.select && initOption.select instanceof SelectData) {
-      // 当select为SelectData时，额外初始化
-      if (initOption.reload == undefined) {
-        initOption.reload = initOption.select.$reload
-      }
-      if (initOption.pagination == undefined && initOption.select.$pagination) {
-        initOption.pagination = initOption.select.$pagination
-      }
-      if (initOption.getData == undefined) {
-        initOption.getData = function(...args) {
-          return (initOption.select as unknown as SelectData).loadData(...args)
-        }
-      }
-    }
     super(initOption, parent, modName)
     this.type = initOption.type
-    this.cascader = initOption.cascader
-    if (this.cascader == undefined) {
-      this.$select = (initOption.select ? (initOption.select instanceof SelectValue ? initOption.select : new SelectValue(initOption.select)) : new SelectValue({}))as unknown as (C extends undefined ? SelectValue<D> : CascaderValue<C, D>)
-    } else {
-      this.$select = (initOption.select ? (initOption.select instanceof CascaderValue ? initOption.select : new CascaderValue(initOption.select as CascaderValueInitOption<C>)) : new CascaderValue({
-        cascader: this.cascader
-      })) as unknown as (C extends undefined ? SelectValue<D> : CascaderValue<C, D>)
-    }
     const option = initOption.option || {}
     const $defaultOption = (this.constructor as typeof SelectEdit).$defaultOption
     this.$option = {
@@ -88,9 +57,6 @@ class SelectEdit<C extends PropertyKey | undefined = undefined, D extends (C ext
         value: undefined,
         ...initOption.search
       }
-    }
-    if (initOption.pagination) {
-      this.$pagination = initOption.pagination instanceof PaginationData ? initOption.pagination : new PaginationData(initOption.pagination)
     }
   }
   $searchData(value?: string) {
@@ -110,12 +76,6 @@ class SelectEdit<C extends PropertyKey | undefined = undefined, D extends (C ext
       }
     } else {
       return Promise.reject({ status: 'fail', msg: '当前选择器不是检索选择器，无法调用$searchData函数！' })
-    }
-  }
-  $clearData() {
-    this.$select.setList([])
-    if (this.$pagination) {
-      this.$pagination.reset(true)
     }
   }
 }